Extract core reporting logic into model
